refactor(kernel): extract INTERCEPT hook dispatch to kernel/dispatch.rs
Phase G.4 of Phase 3 restructure plan.

  - rust/kernel/src/kernel/dispatch.rs (new): impl Kernel block over
    dispatch_native_pre / dispatch_native_post / register_native_hook.
    ~50 LOC.
  - rust/kernel/src/kernel/mod.rs: drops the corresponding methods,
    leaves a "moved to" breadcrumb at the section divider, declares
    `mod dispatch;` alongside `mod locks;` and `mod observability;`.

The HAL-slot accessors (runtime / set_peer_client / peer_client_arc /
set_federation / federation_arc) that lived intermixed in the
dispatch section stay in mod.rs — they're constructor / wiring
state, not hook dispatch.

cargo check -p kernel: clean.
